fix: fetch proofs at specified btc height (#355)
* fix: fetch proofs at specified btc height

* fix: fetch proofs at specified btc height

* fix: publish ss with specified evm block hash

* fix: lint

* fix: lint

* fix: commit chain

* fix: check blockhash in circuit by spv

* fix: constrain header chain size to 1

* fix: constrain header chain size to 1

* fix: constrain header chain size to 1

* fix: constrain header chain size to 1

* fix: move funtion before test

+++ proof-builder-rpc/src/task/state_chain_proof.rs
@@ -87,8 +87,8 @@ async fn spawn_state_chain_ctx_builder(
 
                 let affected = match create_long_running_task(
                     &local_db,
-                    args.start,
-                    args.batch_size,
+                    args.start as i64,
+                    args.batch_size as i64,
                     args.output_proof.clone(),
                     "".to_string(),
                     0,
